// src/api/model/moderationLogResponse.ts
/**
 * Generated by orval v7.18.0 🍺
 * Do not edit manually.
 * Smart Recipes Finder
 * OpenAPI spec version: 2.0.0
 */
import type { ModerationLogResponseRecipeId } from './moderationLogResponseRecipeId';
import type { ModerationLogResponseDraftId } from './moderationLogResponseDraftId';
import type { ModerationLogResponseReason } from './moderationLogResponseReason';
import type { ModerationLogResponseRecipeTitle } from './moderationLogResponseRecipeTitle';
import type { ModerationLogResponseModeratorUsername } from './moderationLogResponseModeratorUsername';

export interface ModerationLogResponse {
  id: number;
  recipe_id?: ModerationLogResponseRecipeId;
  draft_id?: ModerationLogResponseDraftId;
  moderator_id: number;
  action: string;
  reason?: ModerationLogResponseReason;
  recipe_title?: ModerationLogResponseRecipeTitle;
  moderator_username?: ModerationLogResponseModeratorUsername;
  created_at: string;
}