// src/api/model/reportedCommentResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Smart Recipes Finder
 * OpenAPI spec version: 2.0.0
 */
import type { ReportedCommentResponseAuthorUsername } from './reportedCommentResponseAuthorUsername';
import type { ReportedCommentResponseParentCommentId } from './reportedCommentResponseParentCommentId';
import type { ReportedCommentResponseParentContent } from './reportedCommentResponseParentContent';
import type { ReportedCommentResponseParentAuthorUsername } from './reportedCommentResponseParentAuthorUsername';
import type { CommentReportItem } from './commentReportItem';

/**
 * Aggregated view of a commented that has been reported.

Contains the comment itself, all associated reports, and enough
context (recipe, parent comment) for the moderator to decide.
 */
export interface ReportedCommentResponse {
  comment_id: number;
  recipe_id: number;
  recipe_title: string;
  content: string;
  is_deleted: boolean;
  author_id: number;
  author_username?: ReportedCommentResponseAuthorUsername;
  created_at: string;
  parent_comment_id?: ReportedCommentResponseParentCommentId;
  parent_content?: ReportedCommentResponseParentContent;
  parent_author_username?: ReportedCommentResponseParentAuthorUsername;
  report_count: number;
  reports: CommentReportItem[];
}